Login working: Session and messges needs fixing

+++ DoConnect/DocConnectApp/App/Controllers/MessagesController.js
@@ -1,5 +1,5 @@
-﻿app.controller("MessagesController", ["$scope", "MessagesService", "$interval",
-    function ($scope, MessagesService, $interval) {
+﻿app.controller("MessagesController", ["$scope", "MessagesService", "$interval", "$filter", "$ngBootbox",
+    function ($scope, MessagesService, $interval, $filter, $ngBootbox) {
 
         var init_ControlSettings = function () {
             angular.element("#div_Compose_Message").hide();
@@ -36,8 +36,9 @@
 
         //Select All Messages
         $scope.GetAllMessages = function () {
-            MessagesService.GetAllMessages().then
-            (function (result) {
+
+            alert("called GetAllMessages(sessionStorage.Email) " + sessionStorage.Email);
+            MessagesService.GetAllMessages(sessionStorage.Email).then(function (result) {
                 $scope.Messages = result.data;
             });
         };
@@ -55,9 +56,11 @@
             });
         };
 
+        $scope.today = $filter('date')(new Date(), 'yyyy-MM-dd');
+        //alert($filter('time')(new Date(), 'HH:mm:ss'));
         //Insert Message Funtion
-        $scope.NewMessage = function (Sender, Receiver, Subject, Description, Date) {
-            MessagesService.InsertMessage(Sender, Receiver, Subject, Description, Date).success(function () {
+        $scope.NewMessage = function (Receiver, Subject, Description) {
+            MessagesService.InsertMessage(sessionStorage.Email, Receiver, Subject, Description, $scope.today).success(function () {
                 $scope.GetAllMessages();
                 angular.element(".insert").val('');
                 btnSuccess("Message successfully inserted.");
@@ -69,10 +72,13 @@
 
         //Delete Message Funtion
         $scope.DeleteMessage = function () {
-            MessagesService.DeleteMessage($scope.ID).then(function () {
-                $scope.GetAllMessages();
-            }, function (error) {
-                btnAlert("System Error Message", "Delete unsuccessful.");
-            });
+            $ngBootbox.confirm("Are you sure you want to delete this Message?").then(function () {
+                MessagesService.DeleteMessage($scope.ID).then(function () {
+                    $scope.GetAllMessages();
+                    btnSuccess("Message record successfully deleted.");
+                }, function (error) {
+                    btnAlert("System Error Message", "Delete unsuccessful.");
+                });
+            }, function () { });
         };
     }]);
